refactor(cli): many apis are now using IoHost (#40)
### Reason for this change

Changes many internal APIs to use the new IoHost interface.
This will be required so we can accurately track the action for each
log.

Relates to https://github.com/aws/aws-cdk/issues/32292

### Description of changes

Change a lot of logging to use the IoHost. 

There is a temporary annoyance that we also have to pass the action. In
future we will likely replace this with an internal IoHost version that
is action aware. To make this easier for now, a new temporary interface
`IoMessaging` is introduced. This is also why I'm using restructuring `{
ioHost, action }` in many places. To idea being that removing the action
later on will be simpler or at least give cleaner diff.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

existing tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING
GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and
[DESIGN
GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license*