chore: fix jsii publishing in a monorepo (#71)
We were cd'ing into the wrong directories.

The current mechanism builds jsii packages as part of the release step.
This is good because it saves time (they all get built in parallel, and
it's not part of the regular build), but it's bad because if one of them
fails then it'll be hard to fix and retry.

Not sure what to make of this; currently we build jsii packages both at
build time and at release time, which is extra silly. We'll have to pick
one.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>