chore: jsii publishing extracting artifacts to the wrong location (#72)
Follow up to https://github.com/aws/aws-cdk-cli/pull/71.

In contrast to single package jsii publishing, our artifacts don't
contain the entire repo, but rather just the compilation result of a
specific package. This means that upon extraction, it needs to be placed
in the specific package directory instead of the root.

Currently release is failing with:

```console
Error: Expected to find .jsii file in /home/runner/work/aws-cdk-cli/aws-cdk-cli/.repo/packages/@aws-cdk/cli-lib-alpha, but no such file found
```

Which is expected because it operates on an uncompiled repo. Extraction
of the artifact to the correct location will overlay the repo with the
compiled code.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license