chore: give dependabot a deeper path (#106)
Dependabot is complaining it can't find any package version files.

This is the same directory we had in the old repository, so try putting
this back and see if that helps.


---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license