feat: publish CLI plugin contract (#107)
We have already assembled this package, but weren't publishing it just
yet. By publishing the CLI contract, plugin implementors will be able to
take a reliable dependency on the contract types, instead of having to
copy/paste source files.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license