feat(toolkit-lib)!: introducing the AWS CDK Toolkit Library (#108)
The AWS CDK Toolkit Library is a new TypeScript library that allows
integrators to build lifecycle management solutions for CDK
applications. It provides them with full control over all output and
allows to interweave actions with custom steps.

---

This is marked as a breaking change since it is the first release and
want to force `v0.1.0` as the first version.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Co-authored-by: Ian Hou <45278651+iankhou@users.noreply.github.com>