chore(toolkit-lib): rename StackSelector option to `expand` (#109)
In #105 I renamed the interface from `Extend` to `Expand`, but forgot
the options itself and some docs.

This is marked as a chore so it doesn't show up in the changelog for the
first release. That's what we want, because it is not relevant to users
just yet.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license