chore: move minimum Node version back to 14 (#120)
We are currently producing errors when trying to install on Node 14.

Even though we don't know if it works, we didn't officially stop
supporting it yet, so we shouldn't error on old versions.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license