chore: enable `unbound-method` eslint rule (#165)
Unbound method calls using `this` are likely to have unintended effects.

Here is an example with static methods, but the same thing applies to
object methods:

```ts
class Class {
  public static staticMethod() {
    this.otherStaticMethod()
  }

  public static otherStaticMethod() { }
}

// ✅ valid
Class.staticMethod();

// ❌ boom
const x = Class.staticMethod;
x();
```

When assigning a method to a variable, you need to take extra care and
this linter rule is going to remind you.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license