refactor(cli): stack monitor uses io-host (#150)
Provides stack events to the IoHost in a structured way.

The main change here is that `StackActivityMonitor` does not directly
talk to the `ActivityPrinter` anymore. Instead messages are emitted to
the `IoHost`. This caused two related changes:
- Progress tracking has been split into a more generic
`StackProgressMonitor` and progress information is now provided directly
by the `StackActivityMonitor`
- Originally we only had a single timer that covered polling events and
printing them. With the split, we now have two separate timers: one for
polling events and a separate one for updating the view


Removed low-level options:

- `progress` -> now a `CliIoHost` property
- `ci` -> use value from `CliIoHost` directly instead of passing through
all the layers
- `quiet` -> surprisingly this was unused by CLI. We should have
something like this again, but can be implemented in a follow up PR.

Manually tested the changes in addition to integ tests.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license