feat(cdk-assets): release official GA (#172)
Remove the prerelease tag from `cdk-assets`. This tool is ready to be
used for real.

We are still releasing under the `v3-latest` tag, so CDK Pipelines users
won't automatically switch over to it just yet, but this is sending the
right signal about the tool's stability and fitness for purpose.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license