chore: mask bucket name in s3 docs publishing step (#182)
Actually mask the bucket name in s3 docs publishing logs. This isn't a
secret, so it's not bad. The logs for the one run that did accidentally
print the name have already been deleted.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license