chore: change SSM timestamp name to `cli-npm` (#185)
We used to publish under the key name `cli`, but the library publishing
process is writing timestamps under `npm`, `pypi`, etc.

Having the keys

* `npm`
* `pypi`
* `nuget`
* ...
* `cli`

Was confusing, so we're changing it to `cli-npm` instead.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>