feat(bootstrap): Add ECR resource policy for emr-serverless containers (#112)
This PR modifies the default bootstrap template to support EMR
Serverless containers.

Without this policy, containers created using the `DockerImageAsset`
construct cannot be used with EMR Serverless applications.

There is precedence for this pattern to support Lambda function
containers, so I don't think this should be too controversial of a
change.

Relevant documentation on using custom images in EMR Serverless:
-
https://docs.aws.amazon.com/emr/latest/EMR-Serverless-UserGuide/application-custom-image.html

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license