fix(toolkit-lib): `ToolkitError` is not exported (#187)
Fixes `'@aws-cdk/toolkit-lib' has no exported member named
'ToolkitError'`

Because we are now moving parts of the public API into a shared package,
we also need to export the respective types.
During the regular build, type declarations are not bundled. So we need
to take care of it afterwards.
This is not ideal, but required. Thankfully the temporary package is a,
well, temporary construct and we are actively working to get rid of it
(first by adding new stuff, then by moving everything to `toolkit-lib`).

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license