chore(cli): remove unused destroy options (#196)
These are not used downstream, so let's remove them for clarity.

Note that low-level force destroy is a thing in CFN and we _might_
reintroduce this later. It's basically about abandoning failed
resources. But we currently don't support it, so let's not bother.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license