refactor(cli): require approval is a part of the CliIoHost (#151)
`CliIoHost` now governs when to ask for approval or not. This was
previously passed into deploy via deployOptions.

- `deployOptions.requireApproval` was deprecated before, and now it is
no longer being used (this may have to be pulled into a different change
that removes the `requireApproval` option entirely in a minor version
bump of toolkit-lib.
- `CliIoHost` has a property, `requireApproval` that can be set. When
the `requestResponse` API is called, we check to see if the message code
is a known `requireApproval` code and if so, see if the request is
necessary or if the default can be used.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>