feat: record CLI version in manifest package (#184)
Record the current CLI version in the manifest package when we release
the manifest.

This by itself will not trigger a new release of the
`cloud-assembly-schema` package, so the version only gets recorded (and
published!) when there is a new version of `cloud-assembly-schema` to
publish anyway, and the version will only be displayed back to the user
if there is an incompatibility of the major version number.

That means we won't necessarily show the *lowest* possible CLI version
number when the user needs to upgrade, but we'll show a valid CLI
version number that will definitely support the given manifest.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>