refactor(cli): use IoHelper in internal APIs (#209)
This PR is a lot. There is no good way of reviewing it, so thanks for
taking the time.

The main change here is the usage of the `IoHelper` in all internal
APIs. Previously we would pass an `IoHost` and an `action` around as
pair, but separate value. Instead we now use the `IoHelper` which knows
how to send messages with an action.

I also decided to rename all properties and params respectively. This is
partly what makes the PR so large. Note that all changes are to private
APIs.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license