feat(toolkit-lib): emit marker messages during actions (#219)
Introduces the new concept of "Message Spans" to connect multiple
messages for toolkit sub-tasks like "synth" or "build-assets" together.
A span consists of a number of messages that have the same `span` value.

At a minimum, a span is a pair of a `start` and an `end` message. Every
end message also reports the elapsed time since the start of the span.

Message spans will allow integrators to more easily track the flow of
"tasks" completed in the Toolkit.

## Changes

- Refactors the previous `Timer` class to "Message Spans"
- Introduces a new registry for spans at the bottom of the existing
messages registry. A span consists of a start and an end message maker.
- This PR makes use of the new `IoHelper` to provide a nicer API
- There is some more of type magic to ensure correct payloads, I don't
think this can be tested further
- I also changed the message registry to be private again, we are not
ready to have it as part of the public interface (this doesn't affect
our packages)
- Made some improvements to the bundling of declaration files. It's
currently really slow because it ends up scanning through A LOT of SDK
types. Will look into this more in a follow-up PR.

Fixes https://github.com/aws/aws-cdk/issues/33286

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license