chore: remove unused packages (#240)
We moved `cdk-build-tools` and `yarn-cling` over from the old repo, but
they are not really used here.

Just `bockfs` was being used as a testing tool -- move that into the
`aws-cdk` package using it as a test helper.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license