refactor(toolkit): resource metadata (#243)
Pulls out resource-metadata extraction into a separate API so it can
easily be reused in other places.
Also adds support for resource metadata to
`EvaluateCloudFormationTemplate` (internal API) in preparation for a
change in hotswap.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license