fix(cli): hotswap is reporting and running changes that don't happen (#244)
This change is three fold:

- We clean up the private `HotswapOperation` interface (née
`HotswappableChange`) by removing two unused fields and adding in a new
filed in preparation for structured data alongside hotswap messages
- In the hotswap providers, pull gates outside of the apply function.
This prevents hotswaps being reported when nothing would actually be
done. These are local checks and now just run a little bit earlier.
- In `lambda-functions.ts` don't report hotswappable changes for
versions and aliases. These used to be reported with an empty array for
`resourceNames` and a noop `apply` function. This means that literally
nothing happens with these entries, since `apply` doesn't do anything
and the CLI uses `resourceNames` to print anything. Just removing these
as they are reported as part of a change to the function.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license