fix(cli): write notices to stderr or don't write them at all (#221)
(Replaces #188).

On CI systems, the CDK CLI tries to avoid writing to `stderr` because
there are a couple of CI systems that are commonly configured to fail if
any output is written to `stderr`. That means all output, like notices,
must go to `stdout`.

Some commands (like `cdk synth` or `cdk bootstrap --show-template`)
produce usable output on `stdout`, and these are commonly scripted, like
piping their output to a file.

However, because notices must go to `stdout`, these now interfere with
the output of these commands.

This needs a more thorough reworking of the CLI output streams, but
there is a risk of affecting users who are currently relying on the fact
that all output goes to `stdout`.

In this PR, we are doing the first steps to solving this situation:

- Notices will always go to `stderr`, so that they will never interfere
with `stdout` anymore.
- We try to detect what CI system we are running on, and we will
completely suppress notices *unless* we determine that we are running on
a CI system where it is "safe" to write to `sterr` (fail closed).

"Safe" in this case means that the CI system doesn't come with an easy
to toggle checkbox that makes commands fail based on what they print,
instead of their exit codes. The only systems I'm aware of that have
this checkbox are "Azure DevOps", and "TeamCity running PowerShell
scripts".

Even though we know the systems that are "unsafe", we will only show
notices on systems known to be "safe".

Fixes https://github.com/aws/aws-cdk/issues/33589.

Also in this PR, because this grew.

* Introduce `IoDefaultMessages` in the CLI package, which helps migrate
"legacy" logging code to just emit default warning/info/etc messages to
the IoHost.
* Removed the ability to log with a `{ message: 'asdf' }` object to the
global logger functions. This wasn't being used anywhere other than
tests, and it's sort of pointless: if you know the code you should be
using the `MessageMaker` to make a message object; if you don't know the
code you can emit a string. There is no need to look up the right code
given a level and a message object.
* Make it possible for result types to be any type, not just object
types. This is necessary to cover the "result" from legacy logging,
where the result is just a string.
* Updated many tests in a test file (`cli-io-host.test.ts`) that failed
type checking, but succeeded running, and therefore didn't fail the
build of #220.
* Centralized `TestIoHost` into the helper package, and renamed it to
`MockIoHost`.
* Introducing a `FakeIoHost` in the CLI package to assert on messages
emitted to an `IoHost`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>