refactor(cli): re-organize files (#265)
Hopefully the final big code re-org in the CLI before we move code into
the shared package.

This move focuses on moving top-level files into `commands` and `cli`
respectively, to indicate they will stay within the CLI package.
Also removes the `toolkit/error.ts` re-export, which causes many of the
import changes.

Includes a non-consequential bug fix, to the init command and how it's
finding the new `.init-version.json` and
`.recommended-feature-flags.json` files. This was incorrect before and
would happened to work because `init.ts` was in the same directory as
`index.ts`. However because the code is bundled, we need to address the
`init-templates` (and the new files) from the project root location.
This is already done for the templates themselves, but was missed for
`.init-version.json` and `.recommended-feature-flags.json`.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license