feat: @aws-cdk/integ-runner now lives in the CDK CLI repository (#263)
This adds in the code for `@aws-cdk/integ-runner`, which is being
removed from the original `aws-cdk` repository here:
https://github.com/aws/aws-cdk/pull/33835

Process:

* The removal PR should be merged first
* Then this should be merged
* Then we need to find [the latest released
version](https://www.npmjs.com/package/@aws-cdk/integ-runner) and add a
tag named `@aws-cdk/integ-runner@v2.185.0` (or whatever the latest
version is) on the parent commit of this commit.

This proposed to drop the `-alpha` release tag of this package, which
seems fine. We've been running this for years, it hasn't changed much,
and future updates will be easier to consume if we drop the prerelease
tags.

(Stacked PR on top of https://github.com/aws/aws-cdk-cli/pull/246)

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>