refactor(cli): settings doesn't log anymore (#267)
Removes logging from the `Settings` class. Logging was only used when
loading settings from a file and that functionality is currently only
used by the `UserConfiguration` class of the CLI. So I localized the
loading there. If we wanted to re-introduce a generic loading from file
in the future, we just need to build it with the `IoHost` in mind.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license