refactor(cli): tree api with injectable logging (#276)
Updates the `tree.ts` APIs to use an injectable trace function, so that
it can easily be used from old and new code paths. At some point in the
future this should just take an `IoHelper`, but not now.

---
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>